+++ src/global/misc/inc/MResponseMatrixO10.h
@@ -85,7 +85,7 @@ class MResponseMatrixO10 : public MResponseMatrix
            float x9, float x10, float Value = 1);
   void SetMatrix(unsigned int b, MResponseMatrixO9 R9);
 
-  virtual unsigned int GetNBins() const;
+  virtual unsigned long GetNBins() const;
 
   virtual float GetAxisContent(unsigned int b, unsigned int order = 10) const;
   virtual vector<float> GetAxis(unsigned int order = 10) const;
