+++ src/global/misc/inc/MResponseMatrixO12.h
@@ -89,7 +89,7 @@ class MResponseMatrixO12 : public MResponseMatrix
            float x9, float x10, float x11, float x12, float Value = 1);
   void SetMatrix(unsigned int b, MResponseMatrixO11 R11);
 
-  virtual unsigned int GetNBins() const;
+  virtual unsigned long GetNBins() const;
 
   virtual float GetAxisContent(unsigned int b, unsigned int order = 12) const;
   virtual vector<float> GetAxis(unsigned int order = 12) const;
