+++ src/global/misc/inc/MResponseMatrixO7.h
@@ -79,7 +79,7 @@ class MResponseMatrixO7 : public MResponseMatrix
            float x5, float x6, float x7, float Value = 1);
   void SetMatrix(unsigned int b, MResponseMatrixO6 R6);
 
-  virtual unsigned int GetNBins() const;
+  virtual unsigned long GetNBins() const;
 
   virtual float GetAxisContent(unsigned int b, unsigned int order = 7) const;
   virtual vector<float> GetAxis(unsigned int order = 7) const;
