+++ src/global/misc/src/MResponseMatrixO13.cxx
@@ -394,12 +394,12 @@ void MResponseMatrixO13::SetMatrix(unsigned int b, MResponseMatrixO12 R)
 ////////////////////////////////////////////////////////////////////////////////
 
 
-unsigned int MResponseMatrixO13::GetNBins() const
+unsigned long MResponseMatrixO13::GetNBins() const
 {
   // Return the number of bins
 
   if (m_AxisO13.size() < 2) return 0;
-  unsigned int Bins = m_AxisO13.size()-1;
+  unsigned long Bins = m_AxisO13.size()-1;
   return Bins*m_AxesO12[0].GetNBins();
 }
