+++ src/global/misc/src/MResponseMatrixO16.cxx
@@ -407,12 +407,12 @@ void MResponseMatrixO16::SetMatrix(unsigned int b, MResponseMatrixO15 R)
 ////////////////////////////////////////////////////////////////////////////////
 
 
-unsigned int MResponseMatrixO16::GetNBins() const
+unsigned long MResponseMatrixO16::GetNBins() const
 {
   // Return the number of bins
 
   if (m_AxisO16.size() < 2) return 0;
-  unsigned int Bins = m_AxisO16.size()-1;
+  unsigned long Bins = m_AxisO16.size()-1;
   return Bins*m_AxesO15[0].GetNBins();
 }
