+++ src/global/misc/src/MResponseMatrixO4.cxx
@@ -353,12 +353,12 @@ void MResponseMatrixO4::SetMatrix(unsigned int b, MResponseMatrixO3 R)
 ////////////////////////////////////////////////////////////////////////////////
 
 
-unsigned int MResponseMatrixO4::GetNBins() const
+unsigned long MResponseMatrixO4::GetNBins() const
 {
   // Return the number of bins
 
   if (m_AxisO4.size() < 2) return 0;
-  unsigned int Bins = m_AxisO4.size()-1;
+  unsigned long Bins = m_AxisO4.size()-1;
   return Bins*m_AxesO3[0].GetNBins();
 }
