+++ src/global/misc/src/MResponseMatrixO5.cxx
@@ -358,12 +358,12 @@ void MResponseMatrixO5::SetMatrix(unsigned int b, MResponseMatrixO4 R)
 ////////////////////////////////////////////////////////////////////////////////
 
 
-unsigned int MResponseMatrixO5::GetNBins() const
+unsigned long MResponseMatrixO5::GetNBins() const
 {
   // Return the number of bins
 
   if (m_AxisO5.size() < 2) return 0;
-  unsigned int Bins = m_AxisO5.size()-1;
+  unsigned long Bins = m_AxisO5.size()-1;
   return Bins*m_AxesO4[0].GetNBins();
 }
